Info about error: Could not find or load main class org.apache.maven.surefire.booter.ForkedBooter.
The default value for useSystemClassLoader changed between Surefire 2.3 and Surefire 2.4, which was a pretty significant change.
In Surefire 2.3, useSystemClassLoader was false by default, and we used an isolated class loader.
In Surefire 2.4, useSystemClassLoader is true by default.
No value works for everyone, but we think this default is an improvement; a bunch of hard-to-diagnose bugs get better when we useSystemClassLoader=true.

Unfortunately, if useSystemClassLoader is set incorrectly for your application, you're going to have a problem on your hands that can be quite difficult to diagnose.
You might even be forced to read a long documentation page like this one. ;-)

If you're having problems loading classes, try setting useSystemClassLoader=false to see if that helps.
You can do that with the POM snippet below, or by setting -Dsurefire.useSystemClassLoader=false.
If that doesn't work, try setting useSystemClassLoader back to true and setting useManifestOnlyJar to false.

 - https://stackoverflow.com/questions/53010200/maven-surefire-could-not-find-forkedbooter-class
 - https://stackoverflow.com/questions/50661648/spring-boot-fails-to-run-maven-surefire-plugin-classnotfoundexception-org-apache
 - http://maven.apache.org/surefire/maven-surefire-plugin/examples/class-loading.html
 - http://maven.apache.org/surefire/maven-surefire-plugin/test-mojo.html
 - http://maven.apache.org/surefire/maven-failsafe-plugin/examples/class-loading.html
 - http://maven.apache.org/surefire/maven-failsafe-plugin/integration-test-mojo.html